refactor(cli): extract log parser and shared escapeRegex utility
