+++ src/string-utils.ts
@@ -1,0 +1,3 @@
+export function escapeRegex(value: string): string {
+  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+}
